+++ starsky/starsky.foundation.platform/Helpers/FilenamesHelper.cs
@@ -1,5 +1,6 @@
 using System.IO;
 using System.Runtime.InteropServices;
+using System.Text;
 using System.Text.RegularExpressions;
 
 namespace starsky.foundation.platform.Helpers
@@ -15,7 +16,7 @@ namespace starsky.foundation.platform.Helpers
 		/// <returns>Regex object</returns>
 		[GeneratedRegex(
 			"^[a-zA-Z0-9_](?:[a-zA-Z0-9 ._-]*[a-zA-Z0-9])?\\.[a-zA-Z0-9_-]+$",
-			RegexOptions.CultureInvariant,
+			RegexOptions.CultureInvariant | RegexOptions.Singleline,
 			matchTimeoutMilliseconds: 300)]
 		private static partial Regex ValidFileNameRegex();
 
@@ -63,7 +64,7 @@ namespace starsky.foundation.platform.Helpers
 		/// <returns>Regex object</returns>
 		[GeneratedRegex(
 			"\\.[a-zA-Z0-9]{1,4}$",
-			RegexOptions.CultureInvariant,
+			RegexOptions.NonBacktracking,
 			matchTimeoutMilliseconds: 100)]
 		private static partial Regex FileNameWithoutExtensionRegex();
 
@@ -87,7 +88,7 @@ namespace starsky.foundation.platform.Helpers
 		/// <returns>Regex object</returns>
 		[GeneratedRegex(
 			"[^.][a-zA-Z0-9]{1,4}$",
-			RegexOptions.CultureInvariant,
+			RegexOptions.NonBacktracking,
 			matchTimeoutMilliseconds: 100)]
 		private static partial Regex FileExtensionWithoutDotRegex();
 
@@ -108,20 +109,9 @@ namespace starsky.foundation.platform.Helpers
 		}
 
 		/// <summary>
+		/// Return UNIX style parent paths back
 		/// Get Parent Regex
 		/// unescaped regex: /.+(?=\/[^/]+$)/
-		/// pre compiled regex
-		/// Regex.Match
-		/// </summary>
-		/// <returns>Regex object</returns>
-		[GeneratedRegex(
-			".+(?=\\/[^/]+$)",
-			RegexOptions.CultureInvariant,
-			matchTimeoutMilliseconds: 200)]
-		private static partial Regex ParentPathRegex();
-
-		/// <summary>
-		/// Return UNIX style parent paths back
 		/// </summary>
 		/// <param name="filePath">unix style subPath</param>
 		/// <returns>parent folder path</returns>
@@ -131,8 +121,21 @@ namespace starsky.foundation.platform.Helpers
 			{
 				return "/";
 			}
+			
+			var parts = filePath.TrimEnd('/').Split('/');
+			if (parts.Length <= 2)
+			{
+				return "/";
+			}
 
-			var result = ParentPathRegex().Match(filePath).Value;
+			var stringBuilder = new StringBuilder();
+			for (var i = 0; i < parts.Length - 1; i++)
+			{
+				stringBuilder.Append(parts[i] + "/");
+			}
+			
+			var result = stringBuilder.ToString().TrimEnd('/');
+
 			return string.IsNullOrEmpty(result) ? "/" : PathHelper.RemoveLatestSlash(result);
 		}
 	}