+++ starsky/starsky/clientapp/src/shared/url/url-query.ts
@@ -336,7 +336,7 @@ export class UrlQuery {
   };
 
   public UrlApiFeaturesAppSettings = (): string => {
-    return this.prefix + "/api/env/features?v=0.6.0-beta.2";
+    return this.prefix + "/api/env/features?v=0.6.0-rc.1";
   };
 
   public UrlApiDesktopEditorOpenAmountConfirmationChecker = (): string => {
@@ -397,6 +397,11 @@ export class UrlQuery {
 
   public UrlHealthDetails(): string {
     return `${this.prefix}/api/health/details`;
+  }
+
+  public UrlHealthReleaseInfo(v: string): string {
+    if (!v) return `${this.prefix}/api/health/release-info`;
+    return `${this.prefix}/api/health/release-info?v=${v}`;
   }
 
   public UrlHealthCheckForUpdates(): string {